Add files via upload
A)Implement Dynamic implementation of circular queue of integers with following operation:
- Initialize(),insert(), delete(), isempty(), isfull(), display()
							
B) Write a  program to sort elements of a singly linked list in ascending order and display the sorted List.

C) Write a program to create Circular Singly Link list and display it.

D)  Write a  program to read n integers and create two lists such that all positive numbers are in one list and negative numbers are in another list. Display both the lists.

E)Write a  program to count all non-zero elements, odd numbers and even numbers in the singly linked list.	

F) Write a  program which accept the string and check whether the string is Palindrome or not using stack. (Use Static/Dynamic implementation of Stack)